refactor: modularize and separate docker-compose from rancher-compose functions
